+++ lib/git.js
@@ -14,7 +14,9 @@ exports.fetch = function fetch(
       callback(repo, hash);
     });
   } else {
+    console.log("exists");
     deps.git.cwd(`repos/${repo}`).pull(() => {
+      console.log("cdw+pull");
       callback(repo, hash, reporter);
     });
   }
