+++ package.json
@@ -7,8 +7,8 @@
   },
   "main": "server.js",
   "scripts": {
-    "test:lint": "prettier --check server.js 'lib/**/*.js' 'test/**/*.js'",
-    "test:unit": "mocha test/*.js --reporter mocha-multi --reporter-options dot=-,xunit=tested.xml",
+    "test:lint": "prettier --check server.js 'lib/**/*.js' 'tests/**/*.js'",
+    "test:unit": "jest",
     "test": "npm run test:lint && npm run test:unit",
     "test:request": "curl -H \"Content-Type: application/json\" -X POST --data @test/pushEvent.json http://localhost:3000/api",
     "start-dev": "nodemon server.js"
@@ -24,12 +24,9 @@
   },
   "homepage": "https://github.com/Tom-Davidson/complyci#readme",
   "devDependencies": {
-    "expect": "24.1.0",
-    "mocha": "6.0.0",
-    "mocha-multi": "1.0.1",
+    "jest": "24.8.0",
     "nodemon": "1.19.1",
-    "prettier": "1.17.1",
-    "sinon": "7.2.4"
+    "prettier": "1.17.1"
   },
   "dependencies": {
     "co-body": "6.0.0",
